src/ignition/modules/sme_agent/specialized/power_generation_agent.py
#!/usr/bin/env python3
"""Power Generation Specialized Agent for Phase 16.2.

Following crawl_mcp.py methodology for systematic development.
"""

import logging
import time
from typing import Any, Self

from dotenv import load_dotenv

from src.ignition.modules.sme_agent.multi_domain_architecture import (
    AgentTask,
    DomainType,
)

from .base_specialized_agent import BaseSpecializedAgent

load_dotenv()
logger = logging.getLogger(__name__)


class PowerGenerationAgent(BaseSpecializedAgent):
    """Specialized agent for power generation systems."""

    def __init__(self: Self, agent_id: str = "power_generation_agent"):
        super().__init__(
            agent_id=agent_id,
            domain=DomainType.ELECTRICAL,
            industry_type="power_generation",
            max_concurrent_tasks=3,
        )
        self.logger.info(f"Initialized Power Generation Agent: {agent_id}")

    def _initialize_specialized_knowledge(self: Self) -> None:
        self.specialized_knowledge_areas = {
            "thermal_power": {
                "description": "Coal, gas, and oil-fired power plants",
                "keywords": ["thermal", "coal", "gas", "steam", "boiler"],
                "complexity_level": "advanced",
            },
            "renewable_energy": {
                "description": "Solar, wind, and hydroelectric systems",
                "keywords": ["solar", "wind", "hydro", "renewable", "PV"],
                "complexity_level": "standard",
            },
            "grid_integration": {
                "description": "Power grid connection and stability",
                "keywords": ["grid", "transmission", "distribution", "stability"],
                "complexity_level": "advanced",
            },
        }

    def _initialize_regulatory_frameworks(self: Self) -> None:
        self.regulatory_frameworks = [
            "NERC (North American Electric Reliability Corporation)",
            "IEEE Power System Standards",
            "FERC (Federal Energy Regulatory Commission)",
        ]

    def _initialize_process_templates(self: Self) -> None:
        self.process_templates = {
            "thermal_plant": {
                "unit_operations": ["fuel_handling", "combustion", "steam_generation"],
                "critical_parameters": ["efficiency", "emissions", "availability"],
            }
        }

    def _initialize_safety_protocols(self: Self) -> None:
        self.safety_protocols = {
            "electrical_safety": ["High voltage safety", "Arc flash protection"]
        }

    async def process_specialized_task(self: Self, task: AgentTask) -> dict[str, Any]:
        start_time = time.time()

        try:
            return {
                "success": True,
                "analysis": {
                    "task_type": "power_generation",
                    "industry": "Power Generation",
                },
                "processing_time": time.time() - start_time,
            }

        except Exception as e:
            return {
                "success": False,
                "error": f"Processing failed: {e!s}",
                "processing_time": time.time() - start_time,
            }
