pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ign-scripts"
version = "0.2.2"
description = "IGN Scripts - Data processing, automation, and analytics workflows"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0.0",
    "jsonschema>=4.19.0",
    "jinja2>=3.1.0",
    "lxml>=4.9.0",
    "xmltodict>=0.13.0",
    "openpyxl>=3.1.0",
    "structlog>=23.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "click>=8.1.0",
    "rich>=13.5.0",
    "streamlit>=1.28.0",
    "prompt-toolkit>=3.0.0",
    "ruff>=0.11.13",
    "faker>=37.4.0",
    "crawl4ai>=0.6.3",
    "supabase>=2.15.3",
    "mcp>=1.9.4",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pre-commit>=3.5.0",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.12",
]

jupyter = [
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
]

[project.scripts]
ign-scripts = "src.main:main"
ign-scripts-ui = "src.ui.app:main"

[tool.ruff]
target-version = "py311"
line-length = 120

# Force exclusion to ensure documentation files are never checked
force-exclude = true

# ===================================================================
# DOCUMENTATION FILE EXCLUSIONS - Following crawl_mcp.py Methodology
# ===================================================================
# Comprehensive exclusion of markdown and documentation files from
# all linting operations. This prevents wasting time on documentation
# files that don't need Python code quality checks.
#
# Exclusion Strategy:
# 1. Direct file pattern exclusions (*.md, *.txt, etc.)
# 2. Directory-based exclusions (docs/, documentation/)
# 3. Per-file ignore rules with "ALL" to catch any edge cases
# ===================================================================
exclude = [
    "*.md",
    "**/*.md",
    "docs/**/*.md",
    "README.md",
    "CHANGELOG.md",
    "*.markdown",
    "**/*.markdown",
    "docs/",
    "documentation/",
    "*.txt",
    "**/*.txt"
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "ARG001", # unused-function-args
    "ARG002", # unused-method-args
    "ARG003", # unused-class-method-args
    "ARG004", # unused-static-method-args
    "ARG005", # unused-lambda-args
    "D",   # pydocstyle
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "TID", # flake8-tidy-imports
    "Q",   # flake8-quotes
    "UP",  # pyupgrade
    "PT",  # flake8-pytest-style
    "RUF", # Ruff-specific rules
]

ignore = [
    "D100", # Missing docstring in public module
    "D101", # Missing docstring in public class
    "D102", # Missing docstring in public method
    "D103", # Missing docstring in public function
    "D104", # Missing docstring in public package
    "D105", # Missing docstring in magic method
    "D106", # Missing docstring in public nested class
    "D107", # Missing docstring in __init__
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["D", "S101", "ARG", "FBT"]
"scripts/**/*" = ["D", "T201"]
# Comprehensive exclusion of all documentation files from all linting rules
"*.md" = ["ALL"]
"**/*.md" = ["ALL"]
"*.markdown" = ["ALL"]
"**/*.markdown" = ["ALL"]
"*.txt" = ["ALL"]
"**/*.txt" = ["ALL"]
"docs/**/*" = ["ALL"]
"README.*" = ["ALL"]
"CHANGELOG.*" = ["ALL"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "pandas.*",
    "numpy.*",
    "psycopg2.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"
